+++ operator/CHANGELOG.md
@@ -17,6 +17,7 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 
 ### Changed
 
+- Primary proxy reassignment during proxy scaledown, if required, is now enforced (was previously attempted on a best-effort basis).
 - Refactored proxy scaledown handling to improve primary proxy reassignment and node decommissioning with better logging.
 
 ### Removed